Move the NTSTATUS table under windows/errors (Fixes #1213)
windows/nt_status sat beside activedirectory, registry and filesystem, so the
[MS-ERREF] 2.3.1 table was a peer of feature packages rather than of the other
Windows error-code tables. The Win32 system error codes requested in #1208 need
their own package and their own type — NTSTATUS is a bitfield space (Sev/C/N/
Facility/Code) while Win32 codes are flat integers, and the two collide by
value, with 0x00000002 meaning both ERROR_FILE_NOT_FOUND and STATUS_WAIT_2 and
0x00000103 meaning both ERROR_NO_MORE_ITEMS and STATUS_PENDING. 46 of the 1813
NTSTATUS values here fall inside the Win32 code range. HRESULT is a third such
table. All of them belong under one parent.

windows/errors holds no Go files of its own. A package named errors there would
shadow the standard library in every importer, forcing an alias in exactly the
files most likely to need both, and the tables have nothing to share at the type
level: each is a distinct numbering space with a distinct type.

The package name stays nt_status and every exported symbol keeps its name and
behaviour, so this is git mv plus one import line in each of the 44 files that
reference it. Renaming the package to ntstatus, which would fit the convention
elsewhere in the tree, would churn 758 call sites for cosmetics and can be done
on its own later. Restructuring the package internals is deliberately left out
too: the four parallel tables it maintains for one dataset, and the ~1800
exported error values that carry Win32 ERROR_ names over NTSTATUS descriptions,
are follow-up work that only makes sense once the Win32 table exists to compare
against.

Two comments naming the old path are updated so no stale reference remains.
gofmt reordered the import block in eleven files, since the new path sorts
before windows/fileflags where the old one sorted after it.
