+++ network/smb/smb_v10/server/andx_chain_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/header"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/header/flags2"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/types"
+	"github.com/TheManticoreProject/Manticore/windows/errors/nt_status"
 	"github.com/TheManticoreProject/Manticore/windows/fileflags"
-	"github.com/TheManticoreProject/Manticore/windows/nt_status"
 )
 
 // writeAndxRequestDataOffset is where a WriteAndx request's data block begins
