+++ network/smb/smb_v10/server/handler_notify.go
@@ -9,8 +9,8 @@ import (
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/commands/codes"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/commands/command_interface"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/header"
+	"github.com/TheManticoreProject/Manticore/windows/errors/nt_status"
 	"github.com/TheManticoreProject/Manticore/windows/filesystem"
-	"github.com/TheManticoreProject/Manticore/windows/nt_status"
 )
 
 // handleNotifyChange answers NT_TRANSACT_NOTIFY_CHANGE: it tells the client when
