+++ network/smb/smb_v10/server/notify_test.go
@@ -10,9 +10,9 @@ import (
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/commands/codes"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/message/header"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v10/types"
+	"github.com/TheManticoreProject/Manticore/windows/errors/nt_status"
 	"github.com/TheManticoreProject/Manticore/windows/fileflags"
 	"github.com/TheManticoreProject/Manticore/windows/filesystem"
-	"github.com/TheManticoreProject/Manticore/windows/nt_status"
 )
 
 // notifyReply is what a watch's answer carried.
