+++ network/smb/smb_v20/client/session.go
@@ -9,7 +9,7 @@ import (
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v20/message/commands"
 	"github.com/TheManticoreProject/Manticore/network/smb/smb_v20/securitymode"
 	"github.com/TheManticoreProject/Manticore/windows/credentials"
-	"github.com/TheManticoreProject/Manticore/windows/nt_status"
+	"github.com/TheManticoreProject/Manticore/windows/errors/nt_status"
 )
 
 // ntStatusMoreProcessingRequired (STATUS_MORE_PROCESSING_REQUIRED) is the status
